requirements.txt
tk
pillow
ttkthemes